Move functions from series.go to iterable_series.go (no code change)
